--日志系统中短信解绑功能进一步完善
git-svn-id: http://www.ichebao.net/svn/acb/trunk@1336 fbfde4e3-4bc3-4d81-a63b-7b332075e0cf

+++ apps/log/handlers/message.py
@@ -38,13 +38,13 @@ class MessageHandler(BaseHandler):
             data = json_decode(self.request.body)
             sms_type = data.get('sms_type')
             tmobile = data.get('tmobile')
-            umobile = data.get('umobile')
             content = ''
             if check_phone(tmobile) is None:
                 status = ErrorCode.ILLEGAL_MOBILE
                 self.write_ret(status) 
             else:
                 if sms_type == 'JH':
+                    umobile = data.get('umobile')
                     if check_phone(umobile):
                         content = ':SIM' + ' ' + umobile + ':' + tmobile
                         SMSHelper.send_to_terminal(tmobile,content)
@@ -57,20 +57,30 @@ class MessageHandler(BaseHandler):
                     is_clear = data.get('is_clear')
                     ret = SMSHelper.send_to_terminal(tmobile,content)
                     ret = json_decode(ret)
+                    terminal = self.acbdb.get("SELECT id, tid, owner_mobile, login FROM T_TERMINAL_INFO"
+                                               "  WHERE mobile = %s"
+                                               "    AND service_status = 1",
+                                               tmobile)
+                    if not terminal:
+                        status = ErrorCode.TERMINAL_NOT_EXISTED
+                        logging.error("The terminal with tmobile: %s does not exist!", tmobile)
+                        self.write_ret(status)
+                        return
+                    umobile = terminal.owner_mobile
                     if ret['status'] == 0:
                         self.acbdb.execute("UPDATE T_TERMINAL_INFO"
-                                        "  SET service_status = 2"
-                                        "  WHERE mobile = %s",
-                                        tmobile)
+                                            "  SET service_status = 2"
+                                            "  WHERE mobile = %s",
+                                            tmobile)
                         terminals = self.acbdb.query("SELECT id FROM T_TERMINAL_INFO"
-                                                  "  WHERE owner_mobile = %s"
-                                                  "    AND service_status = 1",
-                                                  umobile)
+                                                      "  WHERE owner_mobile = %s"
+                                                      "    AND service_status = 1",
+                                                      umobile)
                         # clear user
                         if len(terminals) == 0:
                             self.acbdb.execute("DELETE FROM T_USER"
-                                            "  WHERE mobile = %s",
-                                            umobile)
+                                                "  WHERE mobile = %s",
+                                                umobile)
                     if is_clear == 1:
                         clear_data(tmobile,self.acbdb)
                     self.write_ret(status)
